+++ sw-sysemu/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added dummy micro-benchmark
 - Added unit test for Perf
 - Added a test sequence generator
+- Extend instruction sequence to add rv64f, rv64a, rv64d
 ### Changed
 - Detect if compiler supports -flto and enable it
 - Change default symbol visibility to "hidden". Export only public interface